requirements.txt
appnope==0.1.4
asttokens==2.4.1
catboost==1.2.5
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
filelock==3.13.4
fonttools==4.51.0
fsspec==2024.3.1
graphviz==0.20.3
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.23.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
liac-arff==2.5.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
plotly==5.21.0
pluggy==1.4.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.2
pytest==8.1.1
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==25.1.2
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
stack-data==0.6.3
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.4.0
torch==2.2.2
tornado==6.4
traitlets==5.14.2
typing_extensions==4.11.0
tzdata==2024.1
wcwidth==0.2.13
xgboost==2.0.3
